+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed provenance column offset calculations for the first line of a file.
 - Fixed cover.lisp coverage displays and simplified the colorization code.
 - Fixed Instance `self` masking on recover from panic in a method.
+- Fixed `:init` inheritance order bug.
 
 ## [1.5.0] - 2026-06-26
 ### Added